chore: refactor some sources
