+++ src/index.ts
@@ -1,2 +1,2 @@
-export { useTus } from './core/useTus';
-export { TusClientProvider } from './core/TusClientProvider';
+export { useTus } from './useTus';
+export { TusClientProvider } from './TusClientProvider';
